Extract event endpoints
